LayoutTests/platform/efl/fast/dom/Element/getBoundingClientRect-expected.txt

Test 1
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "40.000"
PASS rect.width.toFixed(3) is "300.000"
PASS rect.height.toFixed(3) is "100.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 2
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "188.000"
PASS rect.width.toFixed(3) is "320.000"
PASS rect.height.toFixed(3) is "120.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 3
PASS rect.left.toFixed(3) is "18.000"
PASS rect.top.toFixed(3) is "356.000"
PASS rect.width.toFixed(3) is "300.000"
PASS rect.height.toFixed(3) is "100.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 4
PASS rect.left.toFixed(3) is "37.289"
PASS rect.top.toFixed(3) is "483.289"
PASS rect.width.toFixed(3) is "141.421"
PASS rect.height.toFixed(3) is "141.421"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 5
PASS rect.left.toFixed(3) is "8.000"
PASS rect.top.toFixed(3) is "652.000"
PASS rect.width.toFixed(3) is "300.000"
PASS rect.height.toFixed(3) is "374.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 6
Known failure. Skipping.

Test 7
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 1490.000. Was 1506.000.
PASS rect.width.toFixed(3) is "288.000"
FAIL rect.height.toFixed(3) should be 464.000. Was 465.000.
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 8
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2002.000. Was 2018.000.
PASS rect.width.toFixed(3) is "300.000"
PASS rect.height.toFixed(3) is "74.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 9
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2140.000. Was 2156.000.
PASS rect.width.toFixed(3) is "300.000"
PASS rect.height.toFixed(3) is "90.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 10
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2280.000. Was 2296.000.
PASS rect.width.toFixed(3) is "300.000"
PASS rect.height.toFixed(3) is "34.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 11
PASS rect.left.toFixed(3) is "10.000"
FAIL rect.top.toFixed(3) should be 2402.000. Was 2418.000.
PASS rect.width.toFixed(3) is "147.000"
PASS rect.height.toFixed(3) is "34.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 12
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2538.000. Was 2554.000.
PASS rect.width.toFixed(3) is "100.000"
PASS rect.height.toFixed(3) is "100.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 13
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2789.000. Was 2805.000.
PASS rect.width.toFixed(3) is "260.000"
FAIL rect.height.toFixed(3) should be 16.000. Was 17.000.
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 14
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 2869.000. Was 2885.000.
PASS rect.width.toFixed(3) is "300.000"
FAIL rect.height.toFixed(3) should be 232.000. Was 233.000.
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

Test 15
PASS rect.left.toFixed(3) is "8.000"
FAIL rect.top.toFixed(3) should be 3165.000. Was 3181.000.
PASS rect.width.toFixed(3) is "0.000"
PASS rect.height.toFixed(3) is "0.000"
PASS rect.right is rect.left + rect.width
PASS rect.bottom is rect.top + rect.height

PASS successfullyParsed is true

TEST COMPLETE
